210618_Scene_Patch
1. 메인씬 오류 수정
2. 게임오버씬 이미지 추가

+++ Assets/Public/Public_Scripts/AudioManger.cs
@@ -27,7 +27,7 @@ public class AudioManger : MonoBehaviour
             
             PlayerPrefs.SetFloat(BackGroundPref,backgroundValue);
             PlayerPrefs.SetFloat(EffectPref,effectValue);
-            PlayerPrefs.SetFloat(FirstPlay,-1);
+            PlayerPrefs.SetInt(FirstPlay,-1);
         }
         else
         {
